% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchAssignment}
\alias{fetchAssignment}
\title{Access cell-type assignments}
\usage{
fetchAssignment(data)
}
\arguments{
\item{data}{sargent object.}
}
\value{
A named vector of assignments.
}
\description{
Retrieves cell-type assignments in a Sargent Object.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

vec <- fetchAssignment(srgnt)
table(vec)

}
\seealso{
\link{sargentObject} for a sargent object.
}
