% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchAssignmentScore}
\alias{fetchAssignmentScore}
\title{Access assignment scores}
\usage{
fetchAssignmentScore(data, cells = NULL, cell.types = NULL)
}
\arguments{
\item{data}{sargent object.}

\item{cells}{a vector of cells to collect data. Default is all
cells.}

\item{cell.types}{a vector of cell-types to collect data. Default is
all cell-types.}
}
\value{
A \code{data.frame}.
}
\description{
Retrieves assignment scores in a Sargent Object.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

df <- fetchAssignmentScore(srgnt)
head(df)

}
\seealso{
\link{sargentObject} for a sargent object.
}
