% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchCell}
\alias{fetchCell}
\title{Access cells}
\usage{
fetchCell(data, cell.types = NULL, form = c("gather", "spread"))
}
\arguments{
\item{data}{sargent object.}

\item{cell.types}{a vector of cell-types to collect data. Default is
all cell-types.}

\item{form}{indicates the shape of out puts. See \code{details}}
}
\value{
A \code{data.frame}.
}
\description{
Retrieves cells assigned by specific cell-types in a Sargent
               Object.
}
\details{
If \code{form} = \code{gather} returns a \code{data.frame} with two columns
indicating cell ids and associated assignments, respectively. If \code{form}
= \code{spread} returns a \code{data.frame} in which the first column
contains cell ids, next columns contain the associated assignments of each
cell in a binary format (0/1) for given cell-types, and the last column
contains information about the state of the assignemts.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

df <- fetchCell(srgnt, cell.types="TNK")
head(df)

}
\seealso{
\link{sargentObject} for a sargent object.
}
