% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchCelltype}
\alias{fetchCelltype}
\title{Access cell-type assignments}
\usage{
fetchCelltype(data, cells = NULL, form = c("gather", "spread"))
}
\arguments{
\item{data}{sargent object.}

\item{cells}{a vector of cells to collect data. Default is all
cells.}

\item{form}{indicates the shape of out puts. See \code{details}}
}
\value{
A \code{data.frame}.
}
\description{
Retrieves cell-type assignments for specific cells in a
  Sargent Object.
}
\details{
If \code{form} = \code{gather} returns a \code{data.frame} with two
  columns indicating cell ids and associated assignments, respectively. If
  \code{form} = \code{spread} returns a \code{data.frame} in which the first
  column contains cell ids, next columns contain the associated assignments
  of each cell in a binary format (0/1) for given cell-types, and the last
  column contains information about the state of the assignemts.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

cells <- colnames(x$gex)[seq_len(10)]
df <- fetchCelltype(srgnt, cells=cells)
head(df)

}
\seealso{
\link{sargentObject} for a sargent object.
}
