% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchDensityPlot}
\alias{fetchDensityPlot}
\title{Annotation confidency visualization via density-plot}
\usage{
fetchDensityPlot(data)
}
\arguments{
\item{data}{sargent object.}
}
\value{
A ggplot object.
}
\description{
Displays distribution of gini coefficients calculated for each
               cell across cell-types.
}
\details{
A gini coefficient above 0.5 represents good confidency in
assignments. The vertical dashed line represents the threshold used to
identfiy assignemnts with low confidency. If one gene-set is given for
annotation, the gini coefficients will not be calculated and the distribution
of cells' scores are shown instead. In this scenario, the vertical dashed
line represents the threshold used to identfiy cells that belong to the given
gene-set.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)
                           
fetchDensityPlot(srgnt)
                           
}
\seealso{
\link{sargentObject} for a sargent object.
}
