% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchDotPlot}
\alias{fetchDotPlot}
\title{Annotation outcome visualization via dot-plot}
\usage{
fetchDotPlot(data, genes = NULL, min.pct = 0)
}
\arguments{
\item{data}{sargent object.}

\item{genes}{vector of genes to show. Must be chosen from the given
gene-markers.}

\item{min.pct}{the fraction of cells at which to draw the smallest dot
(default is 0).}
}
\value{
A ggplot object.
}
\description{
Displays how gene-markers are expressed across cell-types.
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)

fetchDotPlot(srgnt, 
             genes = c("CD3D","CD3E","IL7R","ANPEP","SLC13A3","SLC17A3",
                       "EMCN","NOTCH4","ADGRL4","CD79A","MS4A1","CD19",
                       "CD14","MS4A7","FCGR3A"), 
             min.pct=0.1)

}
\seealso{
\link{sargentObject} for a sargent object.
}
