% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{fetchMergeSargents}
\alias{fetchMergeSargents}
\title{Merge sargent objects}
\usage{
fetchMergeSargents(..., srgnt.list = NULL)
}
\arguments{
\item{...}{sargent objects to merge.}

\item{srgnt.list}{a list of sargent objects to merge.}
}
\value{
A \code{data.frame} with two columns: 1) cell ids (\code{cell}) and
          cell-type (\code{celltype}).
}
\description{
Merge sargent objects from mutiple level of annotating.
}
\details{
A common strategy to annotate single-cell RNA-seq data is to annotate cells
according to hierarchy of known cell-types. Then, given multiple sargent
objects made in each level of hierarchy, \code{fetchMergeSargents} merges
cell-type assignments based on the cell ids into one column
(\code{celltype}). column \code{celltype} contains assignmets from given
objects separated by ".".
}
\examples{
# getting the data:
x <- get(data("sargentDemoData")) 
# annotating single cell RNA-seq example data 

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)                                   

# get cells annotated as T or Natural Killer cells (TNKs)
cells <- fetchCell(srgnt, cell.types="TNK")$cell

# annotate TNK subset  
srgnt_tnk <- sargentAnnotation(gex=x$gex, cells=cells,
                               gene.sets=x$gene.tnk.sets,
                               adjacent.mtx=x$adjacent.tnk.mtx)
print(srgnt_tnk)

}
\seealso{
\link{sargentObject} for a sargent object.
}
