% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{sargentAnnotation}
\alias{sargentAnnotation}
\title{Annotate scRNA-seq data}
\usage{
sargentAnnotation(
  gex,
  gene.sets,
  gene.sets.neg = NULL,
  cells = NULL,
  gini.min = 0.5,
  gini.sigma = 3,
  adjacent.mtx = NULL,
  n.neighbors = 10,
  score.threshold = NULL,
  only.score = FALSE
)
}
\arguments{
\item{gex}{A numeric matrix of single-cell expression values
where rows are genes and columns are cells. The 
matrix can be stored in sparse format (e.g., 
\code{count} slot from a \code{Seurat} object).}

\item{gene.sets}{A list of gene markers. Each element of the list
is named by the cell type and contains a vector
of associated gene markers. You can also specify
positive and negative weights `[-1, 1]` to each
marker by adding an asterisk (*) sign to genes in
the signature, for example, `0.5*Gene_X`, or 
`-0.75*Gene_Y`.}

\item{gene.sets.neg}{A list of negatvive gene markers. Each element of
the list is named by the cell type and contains a
vector of associated gene markers. Note: this argument
is only available for legacy purposes. Please use
the weighing option in `gene.sets`.}

\item{cells}{Vector of cells to annotate (default is all
cells).}

\item{gini.min}{A numeric scalar, between 0 and 1, specifying the
minimum threshold on the gini index}

\item{gini.sigma}{A numeric scalar specifying the number of sigmas
to use for identifying ambiguous calls (outliers) 
in the distribution of gini indexs.}

\item{adjacent.mtx}{A numeric matrix representing the nearest-neighbor 
graph. Matrix is binary, 1 for neigbors and 0 for
not-neighbors, and square with number of rows and 
coulmns equal to the number of cells.}

\item{n.neighbors}{An intreger specifying the number of nearest
neighbors.}

\item{score.threshold}{A numeric scalar specifiying the minmum threshold 
on scores to annotate cells by a given cell-type. 
Only used if one set of gene-markers is given.}

\item{only.score}{if TRUE, only calculates assignment scores.}
}
\value{
a \link{sargentObject}.
}
\description{
Returns the cell-type of origin for each cell, given a gene 
               expression matrix and sets of gene-markers
}
\examples{
# getting the data:
x <- get(data("sargentDemoData"))

# running cell-type assignment:
srgnt <- sargentAnnotation(gex=x$gex,
                           gene.sets=x$gene.sets,
                           adjacent.mtx=x$adjacent.mtx)
print(srgnt)                                 
table(predicted=fetchAssignment(srgnt), truth=x$mt.data$celltype)

}
\seealso{
\link{sargentObject} for a sargent object.
}
