% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{sargentObject-class}
\alias{sargentObject-class}
\alias{sargentObject}
\alias{print,sargentObject-method}
\alias{sargentObject-method}
\title{S4 class defining sargentObject class}
\usage{
\S4method{print}{sargentObject}(x)
}
\arguments{
\item{x}{Sargent object}
}
\value{
sargentObject
}
\description{
\code{sargentObject} defines sargentAnnotation outputs.
}
\section{Slots}{

\describe{
\item{\code{cells}}{vector of input cell ids.}

\item{\code{cells_type}}{list of annotated cells.}

\item{\code{cells_state}}{vector specifying each cell annotation status:
\itemize{
         \item    \code{classified}: annotated cell. 
         \item    \code{classified-smooth}: annotated cell post-smoothing. 
         \item    \code{unclassified}: unannotated cell. 
         \item    \code{unclassified-smooth}: unannotated cell post-smoothing. 
         }}

\item{\code{cells_score}}{numeric matrix of scores (gene-sets as rows, cells
as columns).}

\item{\code{cells_gini}}{vector specifying gini-score of each cell.}

\item{\code{gini_min}}{minimum threshold on the gini index.}

\item{\code{threshold}}{numeric value specifying the threshold used to
idenfity unannotated cells}

\item{\code{celltype_summary}}{data.frame containing a summary of annotation
                        associated statistics: 
\itemize{ 
          \item    \code{celltype}: given cell-types for annotation. 
          \item    \code{gene}: given markers for annotation. 
          \item    \code{cell_count}: number of cells annotated with given 
                   cell-types. 
          \item    \code{cell_exp_count}: number of cells with detectable 
                   (>0) expression for given genes and cell-types. 
          \item    \code{pct_exp}: fractione of cells with detectable (>0) 
                   expression for given genes and cell-types. 
          \item    \code{mean_exp}: average expression for given genes and 
                   cell-types. 
          \item    \code{z_score}: z-score calculated per gene across 
                   cell-types. 
          }}
}}

