% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{scoreObject-class}
\alias{scoreObject-class}
\alias{scoreObject}
\title{S4 class defining scoreObject class}
\value{
scoreObject
}
\description{
\code{scoreObject} defines sargentAnnotation outputs if `only.score` is TRUE.
}
\section{Slots}{

\describe{
\item{\code{cells_score}}{numeric matrix of scores (gene-sets as rows, cells
as columns).}
}}

